Add Thinkst Canary Console integration
This is the first release of the Thinkst Canary Console integration
into the market place.

A connector is implemented to poll the Canary Console API to get
the latests incidents from the Console. This is then translated into
AlertInfo objects, which will generate cases.

A 'Ping' action is added to check that the API credentials has been
entered correctly, and that the API endpoint is reachable.

A 'Acknowledge Console Alert' action is implemented, this can be used
to acknowledge incidents on the Canary Console.

Signed-off-by: Louis Peens <louis@thinkst.com>