Add Thinkst Canary Console integration (#314)
* Add Thinkst Canary Console integration

This is the first release of the Thinkst Canary Console integration
into the market place.

A connector is implemented to poll the Canary Console API to get
the latests incidents from the Console. This is then translated into
AlertInfo objects, which will generate cases.

A 'Ping' action is added to check that the API credentials has been
entered correctly, and that the API endpoint is reachable.

A 'Acknowledge Console Alert' action is implemented, this can be used
to acknowledge incidents on the Canary Console.

Signed-off-by: Louis Peens <louis@thinkst.com>

* thinkst_canary: Fix return type of fetch_alerts

Fix a small typing issue, the ThinkstConnectorManager.fetch_alerts()
method returns a list of AlertInfo objects, not a dictionary.

Signed-off-by: Louis Peens <louis@thinkst.com>

* thinkst_canary: Add a few basic connector tests

Add a few simple tests to verify that important fields is correctly
filled in when translating Canary Console incidents to GoogleSecOps
alerts. These use mock json data as input, instead of reaching out via
api call.

Signed-off-by: Louis Peens <louis@thinkst.com>

---------

Signed-off-by: Louis Peens <louis@thinkst.com>